backend/capricorn-chat/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-generativeai==0.3.*
flask==2.*
